Replace two of the pots with basic ones without mounting pins
... for the most ridiculous of reasons: it frees up some board space
without altering the physical layout of knobs.

I can't replace all the pots with the basic ones, because the pots
themselves are physically bigger and won't fit, but this odd mix of
potentiometer types works, even if it's all kinds of odd.

Does this board need to be this form factor? No.  I could just move
things around.  But I like the size, I like the bigger 20mm diameter
knobs I'm planning on using, and I'm just doing this ridiculous thing
for fun in the first place, so having a fun form factor is part of it
all.

The ridiculous reason for having four knobs in the first place? That's
how many ADCs the RP2354A has. So this is all very random.

Next up: do a 4-layer stackup and actually start laying things out.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>