rp2354: First stab at pcb layout
This is an insane 6-layer stackup, not because I need six layers
(although it does allow me to be even lazier than usual), but because I
want to use via-in-pad and at that point jlcpcb pretty much ends up
doing the whole precision 6-layer thing anyway.

The minimal RP2354A board makes do with just a simple two-layer stackup,
and I'm sure this could have been done that way too.  But with four
layers it's easy to just use one layer for the +1V1 zone and one for the
+3V3 zone under the RP2354A, and just bring things in with the vias.

This is all connected, but with some ERC noise. Mostly harmless (tm).

Also add JLCPCB part numbers to things, although I need to check them
all several times.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>