1590LB boost / dirt circuit: add LCSC parts numbers
The PMV250 PNP high-side MOSFET is now a AO3401A, and the BC847C is now
a BC849C instead.  Because those were what I went with at JLCPCB.

It's kind of stupid to have this assembled by machine, since it would be
easy to do by hand (admittedly with slightly different size choices),
and this is not a particularly good design for JLCPCB anyway since all
those matched dual transistors are extended parts etc.

But the point was mainly to do PCBA with a simple board so that I can
actually verify that I did it right.  When the RP2354A comes back, it
will take some coding to get it to do anything at all, so if I screwed
up the hw, it's not going to be as obvious (unless I did it really badly
and the magic smoke escapes).

With this, it's going to be pretty obvious when it doesn't work that my
hw design is just wrong.  Or maybe it works fine - it _should_.  I've
successfully tested a very similar basic circuit, but this does have a
different JFET and some other changes.

Anyway, it's sent off.  When I get back home from travels, I should have
something to test.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>