Add an enclosure design for the not-quite-1590LB boost board
The boost board is back, and "works" for some definition of working.

And by that I mean that it is a bit too clean even when actively trying
to mis-bias the signal: I can definitely see it clipping on the
oscilloscope, but it doesn't really change the horrible noises I make on
the guitar very noiticeably.  Except when I crank up the volume, at
which point it makes the horrible noises much much louder.

So it very much works as a boost, and the signal seems to be quite
clean.  I can easily see a 2mV PtP signal on the oscilloscope.  It's all
just _too_ clean to be interesting, despite me trying to mess up that
biasing.

But hey, I printed an enclosure for it.  And it looks very festive with
a green LED and knobs, with the enclosure done in red PLA.  Very
Christmas - this is the sound equivalent of a "ugly xmas sweater" thing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>