Add code directory for the RP2354A board
The rp2354a board is back too, and I soldered the pots and other
through-hole hardware up.  And at least some of it works.

This is literally just a silly 'blink' program that tests all the
potentiometers and the stomp switch, and blinks the activity LED in a
pattern based on them.

That verifies that the board seems to work, although the most important
part - the TAC5112 CODEC and the analog side - is entirely untested and
may be completely broken.  We'll see.

But hey - it does show that my RP2354A itself is alive.  The USB
connector works, the power supplies - both 3.3V and 1.1V - work, and the
12MHz crystal works.

Admittedly that was all just per the minimal board schematic, so it
*should* all have worked and I shouldn't be surprised, but the fact that
it all came alive on the first try is still very heartening.

And the stomp switch acts as a BOOTSEL switch, so programming this thing
is as simple as holding down the stomp switch when plugging in the USB C
connector.  I have a TagConnect SWD setup, but if I never need to use it
I'll be perfectly happy.

Let's see if I can get the TAC5112 to work too.  The initial goal is to
just set up enough of it with i2c that it goes into passthrough mode.

Doing the DMA for the actual I2S data stream is going to be a much more
complicated thing, involving actually learnign how not just the codec
works, but the rp2354 GPIO and DMA engines too.

Baby steps.  But at least the basic rp2354 itself is working.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>