Initial attempt at some random rp2354 PIO program
It's supposed to be a 1.536MHz square wave on pin 8 (BCLK).

It's close enough.

Except for actual i2s I'll need to do FSYNC even with my mono signal,
which means that I'll presumably will need to make the frequency be
3.072MHz: 64 * 48kHz.

But this was my "let's show some life on that pin" work.  I'll have to
think about how I actually want to do that whole FSYNC generation (and
eventually DIN/DOUT too, of course).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>