Do all of BCLK/FSYNC and DIN using rp2354 PIO program
Note that the DIN/DOUT naming is from the standpoint of the TAC5112, so
DIN is input to the codec, and is thus an output for the RP2354.

I fixed my board damage with a soldering iron, so I can look at the
signals again, and it all looks very reasonable.

This feeds just silly pattern data to the PIO engine for the codec
input, since I'm just looking at oscilloscope traces.  And since I
haven't read up on how the DMA engines work yet, it does so manually
with the ARM core just writing to the PIO FIFO.

Also, I realized how stupid I was to do the FSYNC bit manually with
'set' instructions, when it's all entirely static.  Just use 'sideset'
for FSYNC too - exactly like the BCLK output does - which avoids using
any actual instructions for the clocking.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>