Make noises with the TAC5112
Since I now have data input for the DAC, let's change the output source
to be that DAC data.

And what do you know - it works.

I made the test data be a 480Hz sine-wave (100 samples), with POT1
controlling the volume.  Full volume is supposed to be 1V rms - 2.83V
PtP.

According to the oscilloscope I'm actually getting 2.86V PtP output (and
it's really a bit more, the pot multiplier doesn't go to 1), but it does
otherwise look exactly as expected: a proper sine wave with the right
frequency.

The discrepancy is probably just that there's a slight headroom on the
output.

This is all very good.  It is the first actual sign of life from the
TAC5112 outside of responding to the i2c bus, and it means that my i2s
PIO thing is actually working, and apparently I got the bit clocking
right too.

Now, I don't know why the analog bypass didn't work, so it's entirely
possible that I've screwed up the input side to the TAC5112 somehow.

But at least the output side is clearly working, even if it's not
actually useful in this form except as a particularly stupid and limited
signal generator.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>