Drive the i2s bus at a lower frequency of 2.304MHz
It turns out that i2s worked fine at the full 32 bits data width, but
since 24 bits is more than enough for this thing, let's reduce the BCLK
frequency down instead of widening the bit width of the data.

The sample rate remains the same 48kHz, but at least the PIO engine can
run at a lower frequency (that might reduce clock jitter too: the PIO
frequency is all based on the 150MHz sysclk, and while the clock divisor
has a fractional part, the result is still presumably quantized).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>