Start implementing the reading side of i2s
This works to copy the incoming signal to the outgoing side (scaled by
the POT1) although without DMA the software busy loop is just horribly
hacky and the CPU is 100% busy.

But importantly, it means that the input side of the TAC5112 does
actually work.

With this code, the latency seems to be about 800uS, fwiw.  Which sounds
a bit odd, the FIFOs involved on the RP2354 aren't that deep (it
translates to a delay of about 42 samples at 48kHz).  But there are
presumably buffers on the TAC5112 for all the digital processing it does
that I haven't read up on, so ...

Is the result a super-clean signal? No.  I see a fair amount of HF noise
on the oscilloscope, but this "digital passthrough" seems pretty quiet
and accurate with an actual guitar and amp.

The bass response seems fairly bad: looking at the oscilloscope I'd
estimate the -3dB point at around 30Hz.  Which is obviously fine for
guitar, but not if you were to use this for bass.  Again, this might be
partly a TAC5112 programming issue, but I think it's likely at least
partly my AC coupling capacitors.

Increasing the input impedance on the TAC5112 may or may not matter.

Anyway, those are pretty minor details compared to the "input side is
alive and well".  I still think I should figure out why I didn't get the
analog passthrough mode working - it would be good to use that for the
stomp switch - and obviously I need to do DMA and actually make up some
random effect too.

But it now works as a bad guitar pedal with POT1 as an attenuating
volume control.

And "bad pedal" is par for the course for me, so it's unlikely to get
better.  Just hopefully it will eventually get _different_.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>