Make a flanger-like (??) effect
This is likely not what a flanger should sound like, but it certainly
does *something*.  It is very loosely based on looking at what the Daisy
Seed flanger code does, but simplified and without the C++ horrors.

In other words: it's definitely more interesting than the plain volume
control I used to have, but I'm not sure it is a flanger.

And I really should do the whole DMA thing, but I got distracted by
trying to make some half-way interesting effect sound instead.  So it's
still all busy-looping.

This is also when I notice that four knobs are a bit limiting.  I'd like
to also change the LFO frequency, but right now the knobs are volume,
delay, depth and feedback respectively.

The LFO is also kind of simplistic.  It's a triangle wave, but I wonder
if it should be a sine-wave.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>