flanger: fix silly FP errors
Silly cut-and-paste error where the fractional part of the interpolating
array read was done as an integer, which obviously defeated the whole
point of that code.  I think that error was even audible, but maybe
that's just my imagination.

Also, while looking at the generated code to make sure it's not doing
something stupid, I noticed that it's indeed doing something stupid: all
the constants end up forcing double arithmetic, which was against the
whole point.

Now, I could do the mindless thing and write everything as explicit
single-precision (eg "3.14f") but let's just make the compiler do it for
us, and warn if some double aritmetic happens unintentionally.  Because
there is not a reason in the world to ever do doubles here.

That does then cause warnings about the pico-sdk code itself doing that
(probably not intentionally, but I didn't check).  But that warning only
happens when rebuilding the sdk itself.  So it's a slight annoyance, but
only an occasional one.

This makes the generated code look a whole lot saner.

Of course, I'm pretty sure my actual math for the delay is still
entirely broken, even if the effect doesn't sound bad.  And I still want
to make a better LFO, maybe with an option to choose triangle vs sine vs
whatever.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>