Add infrastructure for multiple effects and an example echo/delay
This allows switching between effects with a 1s long-press of the stomp
switch.

To demonstrate, this adds a echo/delay effect (and a quick beep effect
when switching effects).  It's a hacky echo/delay just to demonstrate
the multi-effect thing, but it kind of works.

I really need to more cleanly separate out the delay-line code, right
now it's unnecessarily tied to the flanger code.  And while I did
abstract out the LFO code, I think I need to improve on that abstraction
too so that you can have multiple active LFO's going (and maybe even
chain them into a frequency modulation thing).

That cleanup is a separate thing, but the multi-effect code really makes
it obvious that some of the infrastructure is a bit lacking.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>