Add proper LFO state information model
In order to be able to have multiple independent LFO's, we want to have
a notion of LFO state.  That cleans up the code too.

And to give an example of that, this adds a silly "frequency modulated
signal generator" effect, which ignores the input and just generates a
frequency-modulated signal using two independent LFOs.

I'll clean up the delay line handling next, I think.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>